requirements.txt
mcp==1.6.0
factory-boy==3.3.3
